# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de paiement.",
        "title": "Sikem Payment API",
        "termsOfService": "http://www.sikem.ci/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.sikem.ci/support",
            "email": "support@sikem.ci"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.2"
    },
    "host": "spay.sikem.ci",
    "basePath": "/",
    "paths": {
...
        "/api/services/:id/": {
            "get": {
                "description": "Récuperation des informations de la boutique",
                "tags": [
                    "Services"
                ],
                "summary": "Get service data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.ResServiceAPIGetSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ResFailure"
                        }
                    }
                }
            },
            "put": {
                "description": "Mise à jour des données de la boutique",
                "consumes": [
                    "application/json",
                    "text/xml",
                    "application/x-www-form-urlencoded",
                    "multipart/form-data"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "UpdateInfo service information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pays (ex: civ)",
                        "name": "country",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Une courte description de la boutiuque",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Nom de la boutique",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Site web (ex: https://www.maboutique.ci)",
                        "name": "site_web",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.ResServiceAPIUpdateSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ResFailure"
                        }
                    }
                }
            },
            "delete": {
                "description": "Suppression de la boutique",
                "tags": [
                    "Services"
                ],
                "summary": "Delete service",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.ResServiceAPIDeleteSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ResFailure"
                        }
                    }
                }
            }
        },
        "/api/services/:id/permissions/": {
            "get": {
                "description": "Récuperation des permissions de la boutique",
                "tags": [
                    "Service Permissions"
                ],
                "summary": "Fetch service permissions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.ResServicePermissionAPIFetchSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ResFailure"
                        }
                    }
                }
            }
        },
        "/api/services/:id/permissions/add": {
            "post": {
                "description": "Récuperation des permissions de la boutique",
                "tags": [
                    "Ajout d'un utilisateurs à la boutique"
                ],
                "summary": "Add user to service",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.ResServicePermissionAPIAddSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ResFailure"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.ErrorAPI": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.ProviderModel": {
            "type": "object",
            "required": [
                "description",
                "health_url",
                "name",
                "pay_check_url",
                "pay_url",
                "support_country"
            ],
            "properties": {
                "asynchrone_mode": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "health_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "name_slug": {
                    "type": "string"
                },
                "pay_check_url": {
                    "type": "string"
                },
                "pay_url": {
                    "type": "string"
                },
                "support_country": {
                    "description": "CIV",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TransactionModel"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ResFailure": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ErrorAPI"
                    }
                },
                "is_error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "request_date": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "time_elapsed": {
                    "type": "string"
                }
            }
        },
        "models.ServiceModel": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "current_amount": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "name_slug": {
                    "type": "string"
                },
                "permissions": {
                    "description": "Liste de permission des utilisateurs",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ServicePermissionModel"
                    }
                },
                "site_web": {
                    "type": "string"
                },
                "transactions": {
                    "description": "Liste des transaction",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TransactionModel"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ServicePermissionModel": {
            "type": "object",
            "required": [
                "role"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "role": {
                    "type": "integer"
                },
                "service": {
                    "$ref": "#/definitions/models.ServiceModel"
                },
                "service_id": {
                    "description": "Service",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.UserModel"
                },
                "user_id": {
                    "description": "User",
                    "type": "string"
                }
            }
        },
        "models.TransactionModel": {
            "type": "object",
            "required": [
                "amount",
                "amount_with_fee",
                "currency",
                "operation_mode",
                "operation_state",
                "provider",
                "provider_id",
                "reference_id",
                "service",
                "service_id"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "amount_with_fee": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mode_live": {
                    "type": "boolean"
                },
                "operation_mode": {
                    "description": "CREDIT, DEBIT",
                    "type": "string"
                },
                "operation_msg": {
                    "type": "string"
                },
                "operation_state": {
                    "description": "PENDING, SUCCESS, CANCEL, FAIL",
                    "type": "string"
                },
                "provider": {
                    "$ref": "#/definitions/models.ProviderModel"
                },
                "provider_id": {
                    "description": "Provider",
                    "type": "string"
                },
                "reference_id": {
                    "description": "Réference",
                    "type": "string"
                },
                "service": {
                    "$ref": "#/definitions/models.ServiceModel"
                },
                "service_id": {
                    "description": "Service",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.UserModel": {
            "type": "object",
            "required": [
                "auth_id",
                "first_name",
                "last_name"
            ],
            "properties": {
                "auth_id": {
                    "description": "Identifiant unique keycloak",
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "description": "Information",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "phone_prefix": {
                    "type": "string"
                },
                "role": {
                    "description": "Role",
                    "type": "integer"
                },
                "service_permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ServicePermissionModel"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "services.ResServiceAPIDeleteSuccess": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "deleted": {
                            "type": "boolean"
                        },
                        "service": {
                            "$ref": "#/definitions/models.ServiceModel"
                        }
                    }
                },
                "is_error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "request_date": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "time_elapsed": {
                    "type": "string"
                }
            }
        },
...
        "services.ResServiceAPIGetSuccess": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "service": {
                            "$ref": "#/definitions/models.ServiceModel"
                        }
                    }
                },
                "is_error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "request_date": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "time_elapsed": {
                    "type": "string"
                }
            }
        },
        "services.ResServiceAPIUpdateSuccess": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "service": {
                            "$ref": "#/definitions/models.ServiceModel"
                        }
                    }
                },
                "is_error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "request_date": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "time_elapsed": {
                    "type": "string"
                }
            }
        },
        "services.ResServicePermissionAPIAddSuccess": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "service": {
                            "$ref": "#/definitions/models.ServiceModel"
                        },
                        "servive_permission": {
                            "$ref": "#/definitions/models.ServicePermissionModel"
                        },
                        "user": {
                            "$ref": "#/definitions/models.UserModel"
                        }
                    }
                },
                "is_error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "request_date": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "time_elapsed": {
                    "type": "string"
                }
            }
        },
        "services.ResServicePermissionAPIFetchSuccess": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "service_permissions": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ServicePermissionModel"
                            }
                        }
                    }
                },
                "is_error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "request_date": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "time_elapsed": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}